+++ src/main/java/ofos/entity/UserEntity.java
@@ -1,7 +1,6 @@
 package ofos.entity;
 
 import jakarta.persistence.*;
-import lombok.Builder;
 
 /**
  * UserEntity class is an entity class that maps to the Users table in the database.
@@ -19,7 +18,7 @@ public class UserEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "User_ID")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "Username")
     private String username;
@@ -35,7 +34,7 @@ public class UserEntity {
     public UserEntity() {
     }
 
-    public UserEntity(Long userId, String username, String password, String role, boolean enabled) {
+    public UserEntity(Integer userId, String username, String password, String role, boolean enabled) {
         this.userId = userId;
         this.username = username;
         this.password = password;
@@ -43,11 +42,11 @@ public class UserEntity {
     }
 
     // Getters and Setters
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public void setUserId(Long userId) {
+    public void setUserId(Integer userId) {
         this.userId = userId;
     }
 
@@ -75,7 +74,7 @@ public class UserEntity {
         this.role = role;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return userId;
     }
 }
